feat: move reservation to circulation flow on presentation layer
