+++ app/Presentation/Http/Api/V1/Circulation/Reservation/Register/Controller/RegisterReservationApiController.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Presentation\Http\Api\V1\Reservation\Register\Controller;
+namespace App\Presentation\Http\Api\V1\Circulation\Reservation\Register\Controller;
 
 use App\Application\Circulation\Reservation\UseCases\Register\DTO\RegisterReservationDTO;
 use App\Application\Circulation\Reservation\UseCases\Register\RegisterReservation;
 use App\Application\Identity\User\Exception\UserNotFoundException;
 use App\Application\Inventory\StoredBook\Exception\StoredBookNotFoundException;
 use App\Http\Controllers\Controller;
-use App\Presentation\Http\Api\V1\Reservation\Register\Request\RegisterReservationApiRequest;
-use App\Presentation\Http\Api\V1\Reservation\Shared\Resource\ReservationResource;
+use App\Presentation\Http\Api\V1\Circulation\Reservation\Register\Request\RegisterReservationApiRequest;
+use App\Presentation\Http\Api\V1\Circulation\Reservation\Shared\Resource\ReservationResource;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
