+++ app/Presentation/Http/Api/V1/Circulation/Reservation/Return/Controller/ReturnReservationApiController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Presentation\Http\Api\V1\Reservation\Return\Controller;
+namespace App\Presentation\Http\Api\V1\Circulation\Reservation\Return\Controller;
 
 use App\Application\Circulation\Reservation\Exception\ReservationNotFoundException;
 use App\Application\Circulation\Reservation\UseCases\Return\DTO\ReturnReservationDTO;
@@ -8,8 +8,8 @@ use App\Application\Circulation\Reservation\UseCases\Return\ReturnReservation;
 use App\Domain\Circulation\Reservation\Exception\ReturnReservationDateIsInvalidException;
 use App\Domain\Circulation\Reservation\Status\Exception\InvalidReservationStatusTransitionException;
 use App\Http\Controllers\Controller;
-use App\Presentation\Http\Api\V1\Reservation\Return\Request\ReturnReservationApiRequest;
-use App\Presentation\Http\Api\V1\Reservation\Shared\Resource\ReservationResource;
+use App\Presentation\Http\Api\V1\Circulation\Reservation\Return\Request\ReturnReservationApiRequest;
+use App\Presentation\Http\Api\V1\Circulation\Reservation\Shared\Resource\ReservationResource;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
