+++ app/Presentation/Http/Api/V1/Circulation/Reservation/Shared/Resource/ReservationResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Presentation\Http\Api\V1\Reservation\Shared\Resource;
+namespace App\Presentation\Http\Api\V1\Circulation\Reservation\Shared\Resource;
 
 use App\Domain\Circulation\Reservation\Reservation;
 use Illuminate\Contracts\Support\Arrayable;
